+++ sdk/dataai/Azure.Data.AI/api/Azure.Data.AI.netstandard2.0.cs
@@ -16,9 +16,9 @@ namespace Azure.Data.AI
         public InferenceClient(System.Uri endpoint, Azure.Core.TokenCredential credential, Azure.Data.AI.InferenceClientOptions options) { }
         public virtual Azure.Core.Pipeline.HttpPipeline Pipeline { get { throw null; } }
         public virtual Azure.Response SemanticRerank(Azure.Core.RequestContent content, Azure.RequestContext context = null) { throw null; }
-        public virtual Azure.Response<Azure.Data.AI.SemanticRerankingResult> SemanticRerank(Azure.Data.AI.SemanticRerankingInferenceRequest request, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
+        public virtual Azure.Response<Azure.Data.AI.SemanticRerankingInferenceResult> SemanticRerank(Azure.Data.AI.SemanticRerankingInferenceContent request, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
         public virtual System.Threading.Tasks.Task<Azure.Response> SemanticRerankAsync(Azure.Core.RequestContent content, Azure.RequestContext context = null) { throw null; }
-        public virtual System.Threading.Tasks.Task<Azure.Response<Azure.Data.AI.SemanticRerankingResult>> SemanticRerankAsync(Azure.Data.AI.SemanticRerankingInferenceRequest request, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
+        public virtual System.Threading.Tasks.Task<Azure.Response<Azure.Data.AI.SemanticRerankingInferenceResult>> SemanticRerankAsync(Azure.Data.AI.SemanticRerankingInferenceContent request, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
     }
     public static partial class InferenceClientHostExtensions
     {
@@ -45,9 +45,9 @@ namespace Azure.Data.AI
     public static partial class InferenceModelFactory
     {
         public static Azure.Data.AI.LatencyResult LatencyResult(System.TimeSpan? dataPreprocessDuration = default(System.TimeSpan?), System.TimeSpan? inferenceDuration = default(System.TimeSpan?), System.TimeSpan? postProcessDuration = default(System.TimeSpan?)) { throw null; }
-        public static Azure.Data.AI.SemanticRerankingInferenceRequest SemanticRerankingInferenceRequest(string query = null, System.Collections.Generic.IEnumerable<string> documents = null, bool? returnDocuments = default(bool?), int? topK = default(int?), int? batchSize = default(int?), bool? sort = default(bool?), Azure.Data.AI.SemanticRerankingDocumentType? documentType = default(Azure.Data.AI.SemanticRerankingDocumentType?), string targetPaths = null, string model = null, bool? returnSentenceScore = default(bool?)) { throw null; }
+        public static Azure.Data.AI.SemanticRerankingInferenceContent SemanticRerankingInferenceContent(string query = null, System.Collections.Generic.IEnumerable<string> documents = null, bool? returnDocuments = default(bool?), int? topK = default(int?), int? batchSize = default(int?), bool? sort = default(bool?), Azure.Data.AI.SemanticRerankingDocumentType? documentType = default(Azure.Data.AI.SemanticRerankingDocumentType?), string targetPaths = null, string model = null, bool? returnSentenceScore = default(bool?)) { throw null; }
+        public static Azure.Data.AI.SemanticRerankingInferenceResult SemanticRerankingInferenceResult(System.Collections.Generic.IEnumerable<Azure.Data.AI.SemanticRerankingScore> scores = null, Azure.Data.AI.SemanticRerankingMetaResult meta = null) { throw null; }
         public static Azure.Data.AI.SemanticRerankingMetaResult SemanticRerankingMetaResult(Azure.Data.AI.TokenUsageResult tokenUsage = null, Azure.Data.AI.LatencyResult latency = null, string modelName = null, string modelVersion = null) { throw null; }
-        public static Azure.Data.AI.SemanticRerankingResult SemanticRerankingResult(System.Collections.Generic.IEnumerable<Azure.Data.AI.SemanticRerankingScore> scores = null, Azure.Data.AI.SemanticRerankingMetaResult meta = null) { throw null; }
         public static Azure.Data.AI.SemanticRerankingScore SemanticRerankingScore(int? index = default(int?), string document = null, float? score = default(float?), System.Collections.Generic.IEnumerable<Azure.Data.AI.SentenceScore> sentenceScores = null) { throw null; }
         public static Azure.Data.AI.SentenceScore SentenceScore(int index = 0, float score = 0f) { throw null; }
         public static Azure.Data.AI.TokenUsageResult TokenUsageResult(int? totalTokens = default(int?)) { throw null; }
@@ -85,9 +85,9 @@ namespace Azure.Data.AI
         public static bool operator !=(Azure.Data.AI.SemanticRerankingDocumentType left, Azure.Data.AI.SemanticRerankingDocumentType right) { throw null; }
         public override string ToString() { throw null; }
     }
-    public partial class SemanticRerankingInferenceRequest : System.ClientModel.Primitives.IJsonModel<Azure.Data.AI.SemanticRerankingInferenceRequest>, System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingInferenceRequest>
+    public partial class SemanticRerankingInferenceContent : System.ClientModel.Primitives.IJsonModel<Azure.Data.AI.SemanticRerankingInferenceContent>, System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingInferenceContent>
     {
-        public SemanticRerankingInferenceRequest(string query, System.Collections.Generic.IEnumerable<string> documents) { }
+        public SemanticRerankingInferenceContent(string query, System.Collections.Generic.IEnumerable<string> documents) { }
         public int? BatchSize { get { throw null; } set { } }
         public System.Collections.Generic.IList<string> Documents { get { throw null; } }
         public Azure.Data.AI.SemanticRerankingDocumentType? DocumentType { get { throw null; } set { } }
@@ -98,16 +98,32 @@ namespace Azure.Data.AI
         public bool? Sort { get { throw null; } set { } }
         public string TargetPaths { get { throw null; } set { } }
         public int? TopK { get { throw null; } set { } }
-        protected virtual Azure.Data.AI.SemanticRerankingInferenceRequest JsonModelCreateCore(ref System.Text.Json.Utf8JsonReader reader, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
+        protected virtual Azure.Data.AI.SemanticRerankingInferenceContent JsonModelCreateCore(ref System.Text.Json.Utf8JsonReader reader, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
         protected virtual void JsonModelWriteCore(System.Text.Json.Utf8JsonWriter writer, System.ClientModel.Primitives.ModelReaderWriterOptions options) { }
-        public static implicit operator Azure.Core.RequestContent (Azure.Data.AI.SemanticRerankingInferenceRequest semanticRerankingInferenceRequest) { throw null; }
-        protected virtual Azure.Data.AI.SemanticRerankingInferenceRequest PersistableModelCreateCore(System.BinaryData data, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
+        public static implicit operator Azure.Core.RequestContent (Azure.Data.AI.SemanticRerankingInferenceContent semanticRerankingInferenceContent) { throw null; }
+        protected virtual Azure.Data.AI.SemanticRerankingInferenceContent PersistableModelCreateCore(System.BinaryData data, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
         protected virtual System.BinaryData PersistableModelWriteCore(System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
-        Azure.Data.AI.SemanticRerankingInferenceRequest System.ClientModel.Primitives.IJsonModel<Azure.Data.AI.SemanticRerankingInferenceRequest>.Create(ref System.Text.Json.Utf8JsonReader reader, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
-        void System.ClientModel.Primitives.IJsonModel<Azure.Data.AI.SemanticRerankingInferenceRequest>.Write(System.Text.Json.Utf8JsonWriter writer, System.ClientModel.Primitives.ModelReaderWriterOptions options) { }
-        Azure.Data.AI.SemanticRerankingInferenceRequest System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingInferenceRequest>.Create(System.BinaryData data, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
-        string System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingInferenceRequest>.GetFormatFromOptions(System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
-        System.BinaryData System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingInferenceRequest>.Write(System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
+        Azure.Data.AI.SemanticRerankingInferenceContent System.ClientModel.Primitives.IJsonModel<Azure.Data.AI.SemanticRerankingInferenceContent>.Create(ref System.Text.Json.Utf8JsonReader reader, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
+        void System.ClientModel.Primitives.IJsonModel<Azure.Data.AI.SemanticRerankingInferenceContent>.Write(System.Text.Json.Utf8JsonWriter writer, System.ClientModel.Primitives.ModelReaderWriterOptions options) { }
+        Azure.Data.AI.SemanticRerankingInferenceContent System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingInferenceContent>.Create(System.BinaryData data, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
+        string System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingInferenceContent>.GetFormatFromOptions(System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
+        System.BinaryData System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingInferenceContent>.Write(System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
+    }
+    public partial class SemanticRerankingInferenceResult : System.ClientModel.Primitives.IJsonModel<Azure.Data.AI.SemanticRerankingInferenceResult>, System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingInferenceResult>
+    {
+        internal SemanticRerankingInferenceResult() { }
+        public Azure.Data.AI.SemanticRerankingMetaResult Meta { get { throw null; } }
+        public System.Collections.Generic.IList<Azure.Data.AI.SemanticRerankingScore> Scores { get { throw null; } }
+        protected virtual Azure.Data.AI.SemanticRerankingInferenceResult JsonModelCreateCore(ref System.Text.Json.Utf8JsonReader reader, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
+        protected virtual void JsonModelWriteCore(System.Text.Json.Utf8JsonWriter writer, System.ClientModel.Primitives.ModelReaderWriterOptions options) { }
+        public static explicit operator Azure.Data.AI.SemanticRerankingInferenceResult (Azure.Response response) { throw null; }
+        protected virtual Azure.Data.AI.SemanticRerankingInferenceResult PersistableModelCreateCore(System.BinaryData data, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
+        protected virtual System.BinaryData PersistableModelWriteCore(System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
+        Azure.Data.AI.SemanticRerankingInferenceResult System.ClientModel.Primitives.IJsonModel<Azure.Data.AI.SemanticRerankingInferenceResult>.Create(ref System.Text.Json.Utf8JsonReader reader, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
+        void System.ClientModel.Primitives.IJsonModel<Azure.Data.AI.SemanticRerankingInferenceResult>.Write(System.Text.Json.Utf8JsonWriter writer, System.ClientModel.Primitives.ModelReaderWriterOptions options) { }
+        Azure.Data.AI.SemanticRerankingInferenceResult System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingInferenceResult>.Create(System.BinaryData data, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
+        string System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingInferenceResult>.GetFormatFromOptions(System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
+        System.BinaryData System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingInferenceResult>.Write(System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
     }
     public partial class SemanticRerankingMetaResult : System.ClientModel.Primitives.IJsonModel<Azure.Data.AI.SemanticRerankingMetaResult>, System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingMetaResult>
     {
@@ -125,22 +141,6 @@ namespace Azure.Data.AI
         Azure.Data.AI.SemanticRerankingMetaResult System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingMetaResult>.Create(System.BinaryData data, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
         string System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingMetaResult>.GetFormatFromOptions(System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
         System.BinaryData System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingMetaResult>.Write(System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
-    }
-    public partial class SemanticRerankingResult : System.ClientModel.Primitives.IJsonModel<Azure.Data.AI.SemanticRerankingResult>, System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingResult>
-    {
-        internal SemanticRerankingResult() { }
-        public Azure.Data.AI.SemanticRerankingMetaResult Meta { get { throw null; } }
-        public System.Collections.Generic.IList<Azure.Data.AI.SemanticRerankingScore> Scores { get { throw null; } }
-        protected virtual Azure.Data.AI.SemanticRerankingResult JsonModelCreateCore(ref System.Text.Json.Utf8JsonReader reader, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
-        protected virtual void JsonModelWriteCore(System.Text.Json.Utf8JsonWriter writer, System.ClientModel.Primitives.ModelReaderWriterOptions options) { }
-        public static explicit operator Azure.Data.AI.SemanticRerankingResult (Azure.Response response) { throw null; }
-        protected virtual Azure.Data.AI.SemanticRerankingResult PersistableModelCreateCore(System.BinaryData data, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
-        protected virtual System.BinaryData PersistableModelWriteCore(System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
-        Azure.Data.AI.SemanticRerankingResult System.ClientModel.Primitives.IJsonModel<Azure.Data.AI.SemanticRerankingResult>.Create(ref System.Text.Json.Utf8JsonReader reader, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
-        void System.ClientModel.Primitives.IJsonModel<Azure.Data.AI.SemanticRerankingResult>.Write(System.Text.Json.Utf8JsonWriter writer, System.ClientModel.Primitives.ModelReaderWriterOptions options) { }
-        Azure.Data.AI.SemanticRerankingResult System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingResult>.Create(System.BinaryData data, System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
-        string System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingResult>.GetFormatFromOptions(System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
-        System.BinaryData System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingResult>.Write(System.ClientModel.Primitives.ModelReaderWriterOptions options) { throw null; }
     }
     public partial class SemanticRerankingScore : System.ClientModel.Primitives.IJsonModel<Azure.Data.AI.SemanticRerankingScore>, System.ClientModel.Primitives.IPersistableModel<Azure.Data.AI.SemanticRerankingScore>
     {